refactor: modularize large files and add error boundaries for improved resilience
