+++ backend/app/api/v1/guard/health.py
@@ -1,0 +1,209 @@
+"""
+LLM Guard API — exposes prompt injection scanning as a REST endpoint.
+Copyright (C) 2024 Sarthak Doshi (github.com/SdSarthak)
+SPDX-License-Identifier: AGPL-3.0-only
+
+TODO for contributors (medium difficulty):
+  - Add per-user rate limiting on POST /guard/scan
+  - Persist scan results to the database for audit logs (Completed)
+  - Add a GET /guard/stats endpoint returning block/allow/sanitize counts (Completed)
+"""
+
+import hashlib
+import logging
+from collections import Counter
+from datetime import datetime, timedelta, timezone
+from typing import Optional, TypedDict
+
+from app.api.v1.webhooks import deliver_webhook
+from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query, status
+from fastapi.responses import JSONResponse
+from pydantic import BaseModel
+from sqlalchemy import func
+from sqlalchemy.orm import Session
+
+from app.api.v1.notifications import create_notification
+from app.core.config import settings
+from app.core.database import SessionLocal, get_db
+from app.core.security import get_current_user
+from app.core.rate_limit import guard_scan_rate_limiter
+from app.models.guard_scan_log import GuardScanLog
+from app.models.notification import NotificationType
+from app.models.user import User
+from app.schemas.guard_scan_log import GuardScanLogResponse
+from app.schemas.guard_stats import GuardStatsResponse
+from app.schemas.guard_explain import (
+    ExplainRequest as ExplainRequestModel,
+    ExplainResponse,
+)
+from app.schemas.pagination import PaginatedResponse
+from app.modules.guard import guard_config
+
+router = APIRouter()
+logger = logging.getLogger(__name__)
+
+# Backward-compatible test aliases for the shared rate limiter.
+_scan_attempts_by_user = guard_scan_rate_limiter._local_attempts_by_key
+_RATE_LIMIT_REQUESTS = settings.GUARD_RATE_LIMIT_REQUESTS
+
+
+class ScanRequest(BaseModel):
+    prompt: str
+
+
+class ScanResponse(BaseModel):
+    decision: str
+    confidence: float
+    reasoning: str
+    sanitized_prompt: str | None = None
+    matched_patterns: list[str] = []
+
+
+class GuardConfigRequest(BaseModel):
+    sanitization_level: str
+    malicious_threshold: float
+    suspicious_threshold: float
+
+
+class BulkScanRequest(BaseModel):
+    prompts: list[str]
+
+    def validate_prompts(self) -> None:
+        if not self.prompts:
+            raise ValueError("At least one prompt is required per batch request.")
+
+        if len(self.prompts) > 50:
+            raise ValueError("Maximum 50 prompts allowed per batch request.")
+
+
+class BulkScanResponse(BaseModel):
+    results: list[ScanResponse]
+    total: int
+    processed: int
+
+
+VALID_SANITIZATION_LEVELS = {"low", "medium", "high"}
+
+
+class UserGuardConfig(TypedDict):
+    sanitization_level: str
+    malicious_threshold: float
+    suspicious_threshold: float
+
+
+# Temporary in-memory config store
+user_guard_configs: dict[int, UserGuardConfig] = {}
+
+
+def _infer_detection_type(regex_flag: bool, intent: str) -> str:
+    """Infer whether regex, ML, both, or neither triggered the scan decision."""
+    if not regex_flag and intent == "benign":
+        return "none"
+    if regex_flag and intent == "benign":
+        return "regex"
+    if not regex_flag and intent in {"suspicious", "malicious"}:
+        return "ml"
+    return "combined"
+
+
+def _build_guard_scan_log(user_id: int, prompt: str, result: dict) -> GuardScanLog:
+    """Build a GuardScanLog row without storing raw prompt text."""
+    metadata = result.get("metadata", {})
+    regex_analysis = metadata.get("regex_analysis", {})
+    intent_analysis = metadata.get("intent_analysis", {})
+    decision_reasoning = metadata.get("decision_reasoning", {})
+
+    regex_flag = regex_analysis.get("flag", False)
+    intent = intent_analysis.get("intent", "benign")
+    detection_type = _infer_detection_type(regex_flag, intent)
+
+    return GuardScanLog(
+        user_id=user_id,
+        prompt_hash=hashlib.sha256(prompt.encode()).hexdigest(),
+        decision=result.get("decision", "allow"),
+        confidence=decision_reasoning.get("confidence", 0.0),
+        matched_patterns=regex_analysis.get("matched_patterns", []),
+        detection_type=detection_type,
+        regex_flag=regex_flag,
+        regex_score=regex_analysis.get("risk_score", 0.0),
+        intent=intent,
+        ml_confidence=intent_analysis.get("confidence", 0.0),
+        combined_score=decision_reasoning.get("confidence", 0.0),
+        prompt_length=len(prompt),
+        scanned_at=datetime.utcnow(),
+    )
+
+
+def log_scan(user_id: int, prompt: str, result: dict) -> None:
+    """Log scan details and create block notification without storing raw prompt."""
+    db = SessionLocal()
+
+    try:
+        log = _build_guard_scan_log(user_id, prompt, result)
+
+        db.add(log)
+        db.commit()
+        db.refresh(log)
+
+        if log.decision == "block":
+            create_notification(
+                db=db,
+                user_id=user_id,
+                notification_type=NotificationType.GUARD_BLOCK.value,
+                title="Prompt blocked by LLM Guard",
+                message="A prompt was blocked because it matched high-risk guard rules.",
+                resource_type="guard_scan",
+                resource_id=log.id,
+            )
+            db.commit()
+
+    except Exception:
+        db.rollback()
+        raise
+    finally:
+        db.close()
+
+
+@router.get("/health", tags=["LLM Guard"])
+def guard_health():
+    """Check whether the Guard module is available.
+
+    Returns:
+        A status payload describing the Guard module availability.
+    """
+    return {"module": "llm_guard", "status": "available"}
+
+
+@router.get("/info", tags=["LLM Guard"])
+def guard_info():
+    """Return diagnostic information about the Guard module.
+
+    Returns:
+        A status payload containing device and model details.
+    """
+
+    try:
+        try:
+            import torch
+            device = "cuda" if torch.cuda.is_available() else "cpu"
+        except ImportError:
+            device = "cpu"
+
+        from pathlib import Path
+
+        model_path = Path(guard_config.get_trained_model_path()).name
+
+        return {
+            "module": "llm_guard",
+            "status": "available",
+            "device": device,
+            "model_name": model_path or "pretrained-fallback",
+            "sanitization_level": guard_config.SANITIZATION_LEVEL,
+        }
+    except Exception as e:
+        logger.exception("Failed to get guard info")
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="An error occurred while fetching guard info."
+        )
+
